feat: upgrade alloy crates to 1.4
- alloy: 0.8 → 1.4
- alloy-primitives: 0.8 → 1.0

Breaking API changes addressed:
- FillProvider now takes 3 generics (removed transport type)
- RootProvider parameterized by Network instead of Transport
- with_recommended_fillers() removed (now default behavior)
- on_http() renamed to connect_http()
- Contract call results no longer wrapped in structs
- gas_used changed from u128 to u64

+++ crates/cli/src/commands/deposit.rs
@@ -147,7 +147,7 @@ pub async fn run_v2_deposit(args: &DepositArgs) -> Result<()> {
 }
 
 /// Format gas with thousands separators.
-fn format_gas(gas: u128) -> String {
+fn format_gas(gas: u64) -> String {
     let s = gas.to_string();
     let mut result = String::new();
     for (i, c) in s.chars().rev().enumerate() {
