+++ crates/cli/src/commands/withdraw.rs
@@ -103,7 +103,7 @@ pub async fn run_v2_withdraw(args: &WithdrawArgs) -> Result<()> {
 }
 
 /// Format gas with thousands separators.
-fn format_gas(gas: u128) -> String {
+fn format_gas(gas: u64) -> String {
     let s = gas.to_string();
     let mut result = String::new();
     for (i, c) in s.chars().rev().enumerate() {
